// src/app/feature/model/maxoutsponsor.ts
export class MaxOutSponsor{

    regionId: number;
    centerId: number;
    parishid: number;
    sponsorId: number;
    enrollmentId: number;
    maxoutYear: number;
    regionName: string;
    centerName: string;
    parishName: string;
    sponsorName: string;
    miscAmount: number;
    maxOut: string;
    maxoutMonth: string;
}